Fix login to be compatible with new changes
